Make Common::getIntFromFile() actually return an int
Makes looking at the values returned from linfo's JSON out less wonky, and would remove
the need for consumers of that API needing to cast the values to int

+++ src/Linfo/Common.php
@@ -78,7 +78,7 @@ class Common
     // Get a file who's contents should just be an int. Returns zero on failure.
     public static function getIntFromFile($file)
     {
-        return self::getContents($file, 0);
+        return (int) self::getContents($file, 0);
     }
 
     // Convert bytes to stuff like KB MB GB TB etc
